+++ SandboxiePlus/SandMan/ProxyTunnels/ProxyWindow.h
@@ -1,0 +1,27 @@
+// SPDX-License-Identifier: LicenseRef-Sandboxie-Plus
+// See ../LICENSE for the SandMan license.
+#pragma once
+#include <QDialog>
+#include "ProxyManager.h"
+class QTreeWidget;
+class QLabel;
+class CProxyWindow : public QDialog
+{
+	Q_OBJECT
+public:
+	explicit CProxyWindow(CProxyManager* Manager, QWidget* Parent = nullptr);
+private:
+	void Refresh();
+	void Edit(bool Add);
+	void Import();
+	void Remove();
+	void Run(int Action, bool All);
+	void Assign(bool Detach);
+	void Error(const QString& Message);
+	QStringList Selected() const;
+	CProxyManager* m_Manager;
+	QTreeWidget* m_Profiles;
+	QTreeWidget* m_Boxes;
+	QLabel* m_Notice;
+	QTimer m_Timer;
+};